+++ changelog.md
@@ -1,8 +1,9 @@
-# Version 7.2.2 - dd.mm.2025
+# Version 7.2.2 - 02.02.2025
 
 ## Bugfix
 
-- -
+- Fixed bug in `habapp_rules.actors.energy_save_switch.EnergySaveSwitch` where the max_on_time-timeout switched off the switch, also if external request was active
+- Fixed bug in `habapp_rules.actors.light_bathroom.BathroomLight` which set the night brightness also if the light was switched on via dimming
 
 # Version 7.2.1 - 31.12.2024